rename calculation to nc_statistic
