+++ environment.yml
@@ -30,6 +30,6 @@ dependencies:
 ##############
 # pip install
 - pip:
-#  - -e git+https://github.com/bird-house/eggshell.git#egg=eggshell
+  - -e git+https://github.com/bird-house/eggshell.git#egg=eggshell
   - -e git+https://github.com/NCPP/ocgis#egg=owslib
   - sphinx_rtd_theme
