+++ flyingpigeon/processes/wps_robustness_statistic.py
@@ -8,7 +8,7 @@ from pywps.app.Common import Metadata
 
 from flyingpigeon.utils import extract_archive
 from flyingpigeon.nc_utils import get_variable
-from flyingpigeon.calculation import robustness_stats
+from flyingpigeon.nc_statistic import robustness_stats
 # from flyingpigeon.utils import rename_complexinputs
 # from flyingpigeon.log import init_process_logger
